infra/cloud-function/requirements.txt
google-cloud-securitycenter==1.23.3
google-cloud-pubsub==2.18.4
functions-framework==3.4.0